rename widget file
