+++ packages/pluggableWidgets/gallery-web/src/components/Widget.tsx
